Add DAOSFS FSAL -- New FSAL for DAOS Project
This FSAL allows export of a specially-formatted DAOS(1) container
as a POSIX-like filesystem. The work is very preliminary, but
works well enough as a proof-of-concept. Most basic operations are
supported (directories, file i/o, etc), but some remain unimplemented
due to lack of support in DAOS itself. As DAOS continues to develop,
this FSAL will be updated to take advantage of new capabilities.

1. https://github.com/daos-stack/daos